+++ eras/conway/impl/src/Cardano/Ledger/Conway/Core.hs
@@ -56,7 +56,14 @@ import NoThunks.Class (NoThunks)
 import Numeric.Natural (Natural)
 
 class BabbageEraTxBody era => ConwayEraTxBody era where
+  -- | Lens for getting and setting number of `Coin` that is expected to be in the
+  -- Treasury at the current Epoch
+  currentTreasuryValueTxBodyL :: Lens' (TxBody era) (StrictMaybe Coin)
+
+  -- | Lens for getting and setting `VotingProcedures`.
   votingProceduresTxBodyL :: Lens' (TxBody era) (VotingProcedures era)
+
+  -- | Lens for getting and setting `ProposalProcedures`.
   proposalProceduresTxBodyL :: Lens' (TxBody era) (StrictSeq (ProposalProcedure era))
 
 class BabbageEraPParams era => ConwayEraPParams era where