+++ eras/conway/test-suite/src/Test/Cardano/Ledger/Conway/Translation/TranslatableGen.hs
@@ -62,3 +62,4 @@ genTxBody l = do
     <*> arbitrary
     <*> arbitrary
     <*> arbitrary
+    <*> arbitrary